requirements-python313.txt
# Python 3.13 compatible requirements
# This file excludes packages that may not have wheels for Python 3.13

# Core dependencies
streamlit>=1.31.1
python-dotenv>=1.0.0

# LangChain and OpenAI
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.10
openai>=1.12.0

# PDF processing
PyPDF2>=3.0.1

# Vector database (use ChromaDB instead of FAISS for better Python 3.13 compatibility)
chromadb>=0.4.0

# Google Calendar integration
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.88.0

# HTTP requests
requests>=2.31.0
aiohttp>=3.8.4

# Data processing
numpy>=1.25.0
pandas>=2.0.2
scikit-learn>=1.2.2

# Utilities
tqdm>=4.65.0
rich>=13.4.2
Pygments>=2.15.1
markdown-it-py>=3.0.0
mdurl>=0.1.2

# Development and debugging
watchdog>=3.0.0

# Note: sentencepiece, sentence-transformers, and torch are excluded
# due to Python 3.13 compatibility issues. The application will work
# without these packages, but with reduced text processing capabilities. 